+++ restorers/dataloader/base/__init__.py
@@ -1,2 +1,2 @@
-from .base_dataloader import DatasetFactory
+from .base_dataloader import DatasetFactory, UnsupervisedDatasetFactory
 from .base_low_light_dataloader import LowLightDatasetFactory
